LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/css/css-flexbox/overflow-auto-006-expected.txt


FAIL .flexbox, .inline-flexbox 1 assert_equals:
<div class="flexbox to-be-checked" check-height="" check-accounts-scrollbar="" data-expected-height="47">
  <div class="overflow"><div style="width: 100px; height: 20px"></div></div>
</div>
height expected 47 but got 32
FAIL .flexbox, .inline-flexbox 2 assert_equals:
<div class="flexbox row-reverse to-be-checked" check-height="" check-accounts-scrollbar="" data-expected-height="47">
  <div class="overflow"><div style="width: 100px; height: 20px"></div></div>
</div>
height expected 47 but got 32
FAIL .flexbox, .inline-flexbox 3 assert_equals:
<div class="flexbox to-be-checked" check-height="" check-accounts-scrollbar="" data-expected-height="47">
  <div class="overflow align-self-baseline"><div style="width: 100px; height: 20px"></div></div>
</div>
height expected 47 but got 32
FAIL .flexbox, .inline-flexbox 4 assert_equals:
<div class="flexbox row-reverse to-be-checked" check-height="" check-accounts-scrollbar="" data-expected-height="47">
  <div class="overflow align-self-baseline"><div style="width: 100px; height: 20px"></div></div>
</div>
height expected 47 but got 32
PASS .flexbox, .inline-flexbox 5
PASS .flexbox, .inline-flexbox 6
PASS .flexbox, .inline-flexbox 7
PASS .flexbox, .inline-flexbox 8
PASS .flexbox, .inline-flexbox 9
PASS .flexbox, .inline-flexbox 10
